call super().init_customizations
because of changes in the way Inliner works, init_customizations must call super.

Unfortunately init_customizations in DocUtils calls locals() to get Inliner members in a dict which cause it to fail when called from a subclass. So, to make it work, we must bring super class members to local symbol table. I'd file a PR if docutils was on github.